CLOUD-74241 ldap response fix
