+++ core-api/src/main/java/com/sequenceiq/cloudbreak/api/model/LdapConfigBase.java
@@ -61,6 +61,18 @@ public abstract class LdapConfigBase implements JsonEntity {
     @ApiModelProperty(value = ModelDescriptions.LdapConfigModelDescription.DOMAIN)
     private String domain;
 
+    @NotNull
+    @ApiModelProperty(value = ModelDescriptions.LdapConfigModelDescription.BIND_PASSWORD, required = true)
+    private String bindPassword;
+
+    public String getBindPassword() {
+        return bindPassword;
+    }
+
+    public void setBindPassword(String bindPassword) {
+        this.bindPassword = bindPassword;
+    }
+
     public String getName() {
         return name;
     }
