+++ core-api/src/main/java/com/sequenceiq/cloudbreak/api/model/LdapConfigRequest.java
@@ -1,26 +1,11 @@
 package com.sequenceiq.cloudbreak.api.model;
 
-import javax.validation.constraints.NotNull;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.sequenceiq.cloudbreak.doc.ModelDescriptions;
 
 import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
 
 @ApiModel
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class LdapConfigRequest extends LdapConfigBase {
 
-    @NotNull
-    @ApiModelProperty(value = ModelDescriptions.LdapConfigModelDescription.BIND_PASSWORD, required = true)
-    private String bindPassword;
-
-    public String getBindPassword() {
-        return bindPassword;
-    }
-
-    public void setBindPassword(String bindPassword) {
-        this.bindPassword = bindPassword;
-    }
 }
