+++ core/src/main/java/com/sequenceiq/cloudbreak/converter/LdapConfigToJsonConverter.java
@@ -14,7 +14,7 @@ public class LdapConfigToJsonConverter extends AbstractConversionServiceAwareCon
         json.setDescription(config.getDescription());
         json.setId(config.getId());
         json.setPublicInAccount(config.isPublicInAccount());
-
+        json.setBindPassword(config.getBindPassword());
         json.setBindDn(config.getBindDn());
         json.setServerHost(config.getServerHost());
         json.setServerPort(config.getServerPort());
